pyproject.toml
[project]
name = "livekit-demo"
version = "0.1.0"
description = "LiveKit Voice AI Agent Demo with Braintrust"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents[openai]~=1.2",
    "braintrust[otel]>=0.2.9",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
]
